refactor(runtime/v2): event service
