+++ server/v2/cometbft/streaming.go
@@ -3,8 +3,8 @@ package cometbft
 import (
 	"context"
 
+	"cosmossdk.io/core/event"
 	coreappmgr "cosmossdk.io/server/v2/core/appmanager"
-	"cosmossdk.io/server/v2/core/event"
 	"cosmossdk.io/server/v2/core/store"
 	"cosmossdk.io/server/v2/streaming"
 )
