+++ server/v2/cometbft/utils.go
@@ -22,9 +22,9 @@ import (
 	sdkabci "cosmossdk.io/api/tendermint/abci"
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/comet"
+	"cosmossdk.io/core/event"
 	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/server/v2/core/appmanager"
-	"cosmossdk.io/server/v2/core/event"
 )
 
 // parseQueryRequest parses a RequestQuery into a proto.Message, if it is a proto query
